+++ android/src/main/kotlin/com/deepanshuchaudhary/pdf_manipulator/PdfManipulator.kt
@@ -67,14 +67,14 @@ class PdfManipulator(
         result: MethodChannel.Result,
         sourceFileUri: String?,
         pageCount: Int,
-        byteSize: Int?,
+        byteSize: Number?,
         pageNumbers: List<Int>?,
         pageRanges: List<String>?,
         pageRange: String?,
     ) {
         Log.d(
             LOG_TAG,
-            "splitPdf - IN, sourceFileUri=$sourceFileUri"
+            "splitPdf - IN, sourceFileUri=$sourceFileUri, pageCount=$pageCount, byteSize=$byteSize, pageNumbers=$pageNumbers, pageRanges=$pageRanges, pageRange=$pageRange"
         )
 
         if (!setPendingResult(result)) {
@@ -86,7 +86,10 @@ class PdfManipulator(
         job = uiScope.launch {
             try {
                 val splitPDFPaths: List<String>? = if (byteSize != null) {
-                    getSplitPDFPathsByByteSize(sourceFileUri!!, byteSize.toLong(), activity)
+                    getSplitPDFPathsByByteSize(
+                        sourceFileUri!!,
+                        byteSize.toLong(), activity
+                    )
                 } else if (pageNumbers != null) {
                     getSplitPDFPathsByPageNumbers(sourceFileUri!!, pageNumbers, activity)
                 } else if (pageRanges != null) {
@@ -164,6 +167,166 @@ class PdfManipulator(
             }
         }
         Log.d(LOG_TAG, "removePdfPages - OUT")
+    }
+
+    // For reordering pages of pdf.
+    fun pdfPageReorder(
+        result: MethodChannel.Result,
+        sourceFileUri: String?,
+        pageNumbers: List<Int>?,
+    ) {
+        Log.d(
+            LOG_TAG,
+            "pdfPageReorder - IN, sourceFileUri=$sourceFileUri"
+        )
+
+        if (!setPendingResult(result)) {
+            finishWithAlreadyActiveError(result)
+            return
+        }
+
+        val uiScope = CoroutineScope(Dispatchers.Main)
+        job = uiScope.launch {
+            try {
+                val resultPDFPath: String? =
+                    getPDFPageReorder(sourceFileUri!!, pageNumbers!!, activity)
+
+                finishMergeSuccessfully(resultPDFPath)
+            } catch (e: Exception) {
+                finishWithError(
+                    "pdfPageReorder_exception",
+                    e.stackTraceToString(),
+                    null
+                )
+//            withContext(Dispatchers.IO) {
+//                clearPDFFilesFromCache(context = activity)
+//            }
+            } catch (e: OutOfMemoryError) {
+                finishWithError(
+                    "pdfPageReorder_OutOfMemoryError",
+                    e.stackTraceToString(),
+                    null
+                )
+//            withContext(Dispatchers.IO) {
+//                clearPDFFilesFromCache(context = activity)
+//            }
+            }
+        }
+        Log.d(LOG_TAG, "pdfPageReorder - OUT")
+    }
+
+    // For rotating pages of pdf.
+    fun pdfPageRotator(
+        result: MethodChannel.Result,
+        sourceFileUri: String?,
+        pagesRotationInfo: List<Map<String, Int>>?,
+    ) {
+        Log.d(
+            LOG_TAG,
+            "pdfPageRotator - IN, sourceFileUri=$sourceFileUri"
+        )
+
+        if (!setPendingResult(result)) {
+            finishWithAlreadyActiveError(result)
+            return
+        }
+
+        val uiScope = CoroutineScope(Dispatchers.Main)
+        job = uiScope.launch {
+            try {
+                val newPagesRotationInfo: MutableList<PageRotationInfo> = mutableListOf()
+
+                pagesRotationInfo!!.forEach {
+                    val temp = PageRotationInfo(
+                        pageNumber = it["pageNumber"]!!,
+                        rotationAngle = it["rotationAngle"]!!
+                    )
+                    newPagesRotationInfo.add(temp)
+                }
+
+                val resultPDFPath: String? =
+                    getPDFPageRotator(sourceFileUri!!, newPagesRotationInfo, activity)
+
+                finishMergeSuccessfully(resultPDFPath)
+            } catch (e: Exception) {
+                finishWithError(
+                    "pdfPageRotator_exception",
+                    e.stackTraceToString(),
+                    null
+                )
+//            withContext(Dispatchers.IO) {
+//                clearPDFFilesFromCache(context = activity)
+//            }
+            } catch (e: OutOfMemoryError) {
+                finishWithError(
+                    "pdfPageRotator_OutOfMemoryError",
+                    e.stackTraceToString(),
+                    null
+                )
+//            withContext(Dispatchers.IO) {
+//                clearPDFFilesFromCache(context = activity)
+//            }
+            }
+        }
+        Log.d(LOG_TAG, "pdfPageRotator - OUT")
+    }
+
+    // For reordering, deleting, rotating pages of pdf.
+    fun pdfPageRotatorDeleterReorder(
+        result: MethodChannel.Result,
+        sourceFileUri: String?,
+        pageNumbersForReorder: List<Int>,
+        pageNumbersForDeleter: List<Int>,
+        pagesRotationInfo: List<Map<String, Int>>,
+    ) {
+        Log.d(
+            LOG_TAG,
+            "pdfPageRotatorDeleterReorder - IN, sourceFileUri=$sourceFileUri"
+        )
+
+        if (!setPendingResult(result)) {
+            finishWithAlreadyActiveError(result)
+            return
+        }
+
+        val uiScope = CoroutineScope(Dispatchers.Main)
+        job = uiScope.launch {
+            try {
+                val newPagesRotationInfo: MutableList<PageRotationInfo> = mutableListOf()
+
+                pagesRotationInfo!!.forEach {
+                    val temp = PageRotationInfo(
+                        pageNumber = it["pageNumber"]!!,
+                        rotationAngle = it["rotationAngle"]!!
+                    )
+                    newPagesRotationInfo.add(temp)
+                }
+
+                val resultPDFPath: String? =
+                    getPDFPageRotatorDeleterReorder(sourceFileUri!!,pageNumbersForReorder!!, pageNumbersForDeleter!!,  newPagesRotationInfo, activity)
+
+                finishMergeSuccessfully(resultPDFPath)
+            } catch (e: Exception) {
+                finishWithError(
+                    "pdfPageRotatorDeleterReorder_exception",
+                    e.stackTraceToString(),
+                    null
+                )
+//            withContext(Dispatchers.IO) {
+//                clearPDFFilesFromCache(context = activity)
+//            }
+            } catch (e: OutOfMemoryError) {
+                finishWithError(
+                    "pdfPageRotatorDeleterReorder_OutOfMemoryError",
+                    e.stackTraceToString(),
+                    null
+                )
+//            withContext(Dispatchers.IO) {
+//                clearPDFFilesFromCache(context = activity)
+//            }
+            }
+        }
+        Log.d(LOG_TAG, "pdfPageRotatorDeleterReorder - OUT")
     }
 
     fun cancelManipulations(