+++ lib/pdf_manipulator.dart
@@ -1,4 +1,11 @@
 export 'src/pdf_manipulator.dart' show PdfManipulator;
 
 export 'src/pdf_manipulator_method_channel.dart'
-    show PDFMergerParams, PDFSplitterParams, PDFPageDeleterParams;
+    show
+        PDFMergerParams,
+        PDFSplitterParams,
+        PDFPageDeleterParams,
+        PDFPageReorderParams,
+        PDFPageRotatorParams,
+        PageRotationInfo,
+        PDFPageRotatorDeleterReorderParams;