+++ lib/src/pdf_manipulator.dart
@@ -20,10 +20,36 @@ class PdfManipulator {
 
   /// Deletes pages from provided pdf file.
   ///
-  /// Returns the path or uri of the resultant merged file or null if operation was cancelled.
+  /// Returns the path or uri of the resultant file or null if operation was cancelled.
   /// Throws exception on error.
   Future<String?> pdfPageDeleter({PDFPageDeleterParams? params}) {
     return PdfManipulatorPlatform.instance.pdfPageDeleter(params: params);
+  }
+
+  /// Reorders pages of provided pdf file.
+  ///
+  /// Returns the path or uri of the resultant file or null if operation was cancelled.
+  /// Throws exception on error.
+  Future<String?> pdfPageReorder({PDFPageReorderParams? params}) {
+    return PdfManipulatorPlatform.instance.pdfPageReorder(params: params);
+  }
+
+  /// Rotate pages of provided pdf file.
+  ///
+  /// Returns the path or uri of the resultant file or null if operation was cancelled.
+  /// Throws exception on error.
+  Future<String?> pdfPageRotator({PDFPageRotatorParams? params}) {
+    return PdfManipulatorPlatform.instance.pdfPageRotator(params: params);
+  }
+
+  /// Rotate, delete, reorder pages of provided pdf file.
+  ///
+  /// Returns the path or uri of the resultant file or null if operation was cancelled.
+  /// Throws exception on error.
+  Future<String?> pdfPageRotatorDeleterReorder(
+      {PDFPageRotatorDeleterReorderParams? params}) {
+    return PdfManipulatorPlatform.instance
+        .pdfPageRotatorDeleterReorder(params: params);
   }
 
   /// Cancels running manipulations.